Fix authentication bugs

+++ db/migrate/20191025004638_add_balance_column.rb
@@ -1,5 +1,5 @@
 class AddBalanceColumn < ActiveRecord::Migration[5.1]
   def change
-    add_column(:users, :balance, :decimal)
+    add_column(:users, :balance, :float)
   end
 end
